accesskit_unix::atspi::interfaces::accessible

Struct RootAccessibleInterface

pub(crate) struct RootAccessibleInterface {
    bus_name: OwnedUniqueName,
    root: PlatformRoot,
}

Fields
bus_name: OwnedUniqueNameroot: PlatformRoot
Implementations

impl RootAccessibleInterface

pub fn new(bus_name: OwnedUniqueName, root: PlatformRoot) -> Self

impl RootAccessibleInterface

fn name(&self) -> Result<String>

fn description(&self) -> &str

fn parent(&self) -> OwnedObjectAddress

fn child_count(&self) -> Result<i32>

fn locale(&self) -> &str

fn accessible_id(&self) -> &str

fn get_child_at_index(&self, index: i32) -> Result<(OwnedObjectAddress,)>

fn get_children(&self) -> Result<Vec<OwnedObjectAddress>>

fn get_index_in_parent(&self) -> i32

fn get_relation_set(&self) -> Vec<(RelationType, Vec<OwnedObjectAddress>)>

fn get_role(&self) -> Role

fn get_state(&self) -> StateSet

fn get_application(&self) -> (OwnedObjectAddress,)

fn get_interfaces(&self) -> InterfaceSet

impl RootAccessibleInterface

pub async fn name_changed( &self, __zbus__signal_emitter: &SignalEmitter<'_>, ) -> Result<()>

Emit the “PropertiesChanged” signal with the new value for the Name property.

This method should be called if a property value changes outside its setter method.

pub async fn description_changed( &self, __zbus__signal_emitter: &SignalEmitter<'_>, ) -> Result<()>

Emit the “PropertiesChanged” signal with the new value for the Description property.

This method should be called if a property value changes outside its setter method.

pub async fn parent_changed( &self, __zbus__signal_emitter: &SignalEmitter<'_>, ) -> Result<()>

Emit the “PropertiesChanged” signal with the new value for the Parent property.

This method should be called if a property value changes outside its setter method.

pub async fn child_count_changed( &self, __zbus__signal_emitter: &SignalEmitter<'_>, ) -> Result<()>

Emit the “PropertiesChanged” signal with the new value for the ChildCount property.

This method should be called if a property value changes outside its setter method.

pub async fn locale_changed( &self, __zbus__signal_emitter: &SignalEmitter<'_>, ) -> Result<()>

Emit the “PropertiesChanged” signal with the new value for the Locale property.

This method should be called if a property value changes outside its setter method.

pub async fn accessible_id_changed( &self, __zbus__signal_emitter: &SignalEmitter<'_>, ) -> Result<()>

Emit the “PropertiesChanged” signal with the new value for the AccessibleId property.

This method should be called if a property value changes outside its setter method.